refactor(core): refactor nrf, disable uart in booloader to save flash space
[no changelog]
