+++ core/SConscript.kernel
@@ -46,6 +46,7 @@ FEATURES_WANTED = [
     "secmon_layout",
     "tropic",
     "usb",
+    "smp",
     ]
 
 if DISABLE_OPTIGA:
@@ -262,10 +263,6 @@ env = Environment(
 )
 
 FEATURES_AVAILABLE = models.configure_board(TREZOR_MODEL, HW_REVISION, FEATURES_WANTED, env, CPPDEFINES_HAL, SOURCE_HAL, PATH_HAL)
-
-if 'nrf' in FEATURES_AVAILABLE:
-    FEATURES_AVAILABLE.append('smp')
-    CPPDEFINES_HAL.append('USE_SMP')
 
 FEATURE_FLAGS["AES_GCM"] = FEATURE_FLAGS["AES_GCM"] or "tropic" in FEATURES_AVAILABLE
